# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Halal Grocery Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all active stores
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreListResponse"
...
  /stores/{storeSlug}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store by slug
      parameters:
        - name: storeSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreDetail"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{storeSlug}/products:
    get:
      operationId: getStoreProducts
      tags: [stores]
      summary: Get products for a store
      parameters:
        - name: storeSlug
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: productType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 40
      responses:
        "200":
          description: Store products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /stores/{storeSlug}/reviews:
    get:
      operationId: getStoreReviews
      tags: [stores]
      summary: Get reviews for a store
      parameters:
        - name: storeSlug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Store reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
  /stores/{storeSlug}/pickup-slots:
    get:
      operationId: getPickupSlots
      tags: [stores]
      summary: Get available pickup slots for a store
      parameters:
        - name: storeSlug
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Available pickup slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PickupSlot"

  # PRODUCTS
...
  /admin/stores:
    get:
      operationId: adminListStores
      tags: [stores]
      summary: Admin - list all stores
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - error
        - message
...
    Store:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        logo:
          type: string
          nullable: true
        banner:
          type: string
          nullable: true
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        lat:
          type: number
        lng:
          type: number
        isHalalCertified:
          type: boolean
        halalCertNumber:
          type: string
          nullable: true
        rating:
          type: number
        totalRatings:
          type: integer
        avgPrepTimeMinutes:
          type: integer
        deliveryAvailable:
          type: boolean
        pickupAvailable:
          type: boolean
        curbsideAvailable:
          type: boolean
        minOrderAmount:
          type: number
        convenienceFee:
          type: number
        openingHoursJson:
          type: object
        isMasjidAffiliated:
          type: boolean
        masjidName:
          type: string
          nullable: true
        isActive:
          type: boolean
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - phone
        - address
        - city
        - province
        - postalCode
        - lat
        - lng
        - isHalalCertified
        - rating
        - totalRatings
        - avgPrepTimeMinutes
        - deliveryAvailable
        - pickupAvailable
        - curbsideAvailable
        - minOrderAmount
        - convenienceFee
        - openingHoursJson
        - isMasjidAffiliated
        - isActive
        - isApproved
        - createdAt
    StoreDetail:
      allOf:
        - $ref: "#/components/schemas/Store"
        - type: object
          properties:
            productsCount:
              type: integer
            recentReviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    StoreListResponse:
      type: object
      properties:
        stores:
          type: array
          items:
            $ref: "#/components/schemas/Store"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - stores
        - total
        - page
        - totalPages

    # PRODUCT SCHEMAS
    Product:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        productType:
          type: string
          enum: [PACKAGED, FRESH_MEAT, PRODUCE, FROZEN, SPICES, BAKERY, DAIRY, BEVERAGES]
        category:
          type: string
        subcategory:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        price:
          type: number
        comparePrice:
          type: number
          nullable: true
        unit:
          type: string
        stockQty:
          type: number
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        isHalalCertified:
          type: boolean
        isFreshMeat:
          type: boolean
        meatAnimalType:
          type: string
          nullable: true
        estimatedWeightKg:
          type: number
          nullable: true
        pricePerKg:
          type: number
          nullable: true
        tags:
          type: array
          items:
            type: string
        freshnessLabel:
          type: string
          nullable: true
        store:
          $ref: "#/components/schemas/Store"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - slug
        - name
        - productType
        - category
        - images
        - price
        - unit
        - stockQty
        - isActive
        - isFeatured
        - isHalalCertified
        - isFreshMeat
        - tags
        - createdAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
...
    Review:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        storeId:
          type: string
        orderId:
          type: string
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        reply:
          type: string
          nullable: true
        isPublic:
          type: boolean
        user:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            avatar:
              type: string
              nullable: true
          required:
            - firstName
            - lastName
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - storeId
        - orderId
        - rating
        - isPublic
        - createdAt
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        averageRating:
          type: number
        ratingBreakdown:
          type: object
          properties:
            "1":
              type: integer
            "2":
              type: integer
            "3":
              type: integer
            "4":
              type: integer
            "5":
              type: integer
          required: ["1", "2", "3", "4", "5"]
      required:
        - reviews
        - total
        - averageRating
        - ratingBreakdown

    # PICKUP SLOT
    PickupSlot:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        date:
          type: string
          format: date-time
        startTime:
          type: string
        endTime:
          type: string
        maxOrders:
          type: integer
        bookedOrders:
          type: integer
        isAvailable:
          type: boolean
        spotsLeft:
          type: integer
      required:
        - id
        - storeId
        - date
        - startTime
        - endTime
        - maxOrders
        - bookedOrders
        - isAvailable
        - spotsLeft

    # LOYALTY
...